Move serviceWorkerHelper to separate module
The naming of this is all a bit questionable — perhaps some renaming is
in order later.
